#8 Спринт 12 - Финализировать проект: -Исправлены ошибки синтаксиса; -Добавлены методы валидации полей реквестов создания и обновления; -Добавлены правки и класс костыль для прохождения тестов postman; -Добавлены тесты на все публичные методы хранилища; -Исправлен codeStyle.

+++ src/main/java/ru/yandex/practicum/filmorate/storage/mapper/GenreRowMapper.java
@@ -12,7 +12,7 @@ public class GenreRowMapper implements RowMapper<GenreDTO> {
     @Override
     public GenreDTO mapRow(ResultSet resultSet, int rowNum) throws SQLException {
         GenreDTO genre = new GenreDTO();
-        genre.setId(resultSet.getLong("id"));
+        genre.setId(resultSet.getInt("id"));
         genre.setName(resultSet.getString("genre"));
         return genre;
     }
